Move domain unit tests to jest
