+++ package.json
@@ -34,6 +34,7 @@
     "grunt-zip": "~0.15.0",
     "http-server": "~0.9.0",
     "husky": "^1.3.1",
+    "jest": "^24.1.0",
     "lint-staged": "^8.1.4",
     "load-grunt-tasks": "~0.6.0",
     "prettier": "^1.16.4",
@@ -67,7 +68,8 @@
     ],
     "env": {
       "browser": true,
-      "qunit": true
+      "qunit": true,
+      "jest": true
     },
     "globals": {
       "sigma": "writable"
@@ -82,7 +84,16 @@
       "no-bitwise": "off"
     }
   },
+  "babel": {
+    "presets": [
+      "@babel/preset-env"
+    ]
+  },
   "dependencies": {
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.3.3",
+    "@babel/preset-env": "^7.3.1",
+    "babel-jest": "^24.1.0",
     "conrad": "^0.1.0",
     "eslint-config-prettier": "^4.0.0",
     "qunit": "^2.9.1"
