--- test/unit.graph.js
@@ -1,756 +1,0 @@
-QUnit.module("sigma.classes.graph");
-
-QUnit.test("Basic manipulation", assert => {
-  const opts = {};
-  const settings = new sigma.classes.configurable(opts);
-  const graph = {
-    nodes: [
-      {
-        id: "n0",
-        label: "Node 0",
-        myNodeAttr: 123
-      },
-      {
-        id: "n1",
-        label: "Node 1"
-      },
-      {
-        id: "n2",
-        label: "Node 2"
-      },
-      {
-        id: "n3",
-        label: "Node 3"
-      }
-    ],
-    edges: [
-      {
-        id: "e0",
-        source: "n0",
-        target: "n1",
-        myEdgeAttr: 123
-      },
-      {
-        id: "e1",
-        source: "n1",
-        target: "n2"
-      },
-      {
-        id: "e2",
-        source: "n1",
-        target: "n3"
-      },
-      {
-        id: "e3",
-        source: "n2",
-        target: "n3"
-      },
-      {
-        id: "e4",
-        source: "n2",
-        target: "n2"
-      }
-    ]
-  };
-
-  // Initialize the graph:
-  let myGraph = new sigma.classes.graph(settings);
-
-  opts.immutable = opts.clone = true;
-  myGraph.addNode(graph.nodes[0]);
-  opts.clone = false;
-  myGraph.addNode(graph.nodes[1]);
-  myGraph.addNode(graph.nodes[2]);
-  myGraph.addNode(graph.nodes[3]);
-
-  opts.immutable = opts.clone = true;
-  myGraph.addEdge(graph.edges[0]);
-  opts.clone = false;
-  myGraph.addEdge(graph.edges[1]);
-  myGraph.addEdge(graph.edges[2]);
-  myGraph.addEdge(graph.edges[3]);
-  myGraph.addEdge(graph.edges[4]);
-
-  // NODES:
-  // ******
-  assert.deepEqual(
-    graph.nodes[0],
-    myGraph.nodes(graph.nodes[0].id),
-    '"addNode" works and the node properties have been preserved.'
-  );
-
-  assert.notStrictEqual(
-    graph.nodes[0],
-    myGraph.nodes(graph.nodes[0].id),
-    'With {clone: true}, "addNode" creates a new object.'
-  );
-
-  assert.equal(
-    graph.nodes[1],
-    myGraph.nodes(graph.nodes[1].id),
-    'With {clone: false}, "addNode" keeps the same object.'
-  );
-
-  myGraph.nodes(graph.nodes[0].id).id = "new_n0";
-  assert.strictEqual(
-    graph.nodes[0].id,
-    myGraph.nodes(graph.nodes[0].id).id,
-    "With {immutable: true}, node ids in the graph are not writable."
-  );
-
-  const node = myGraph.nodes(graph.nodes[1].id);
-  node.id = "new_n0";
-  assert.strictEqual(
-    "new_n0",
-    node.id,
-    "With {immutable: false}, node ids in the graph are writable."
-  );
-  node.id = "n1";
-
-  myGraph.nodes(graph.nodes[0].id).label = "New node 0";
-  assert.strictEqual(
-    "New node 0",
-    myGraph.nodes(graph.nodes[0].id).label,
-    "Other node attributes are writable."
-  );
-  myGraph.nodes(graph.nodes[0].id).label = "Node 0";
-
-  assert.notStrictEqual(
-    myGraph.nodes(),
-    myGraph.nodes(),
-    '"nodes" without arguments returns a copy of the nodes array.'
-  );
-
-  assert.strictEqual(
-    myGraph.nodes("unexisting_id"),
-    undefined,
-    '"nodes" with an unreferenced id returns undefined and does not throw an error.'
-  );
-
-  assert.deepEqual(
-    myGraph.nodes(["n0", "n1", "n0"]),
-    [graph.nodes[0], graph.nodes[1], graph.nodes[0]],
-    '"nodes" with a strings array as arguments returns the array of specified nodes.'
-  );
-
-  assert.throws(
-    () => myGraph.nodes(["n0", "n1", {}]),
-    /nodes: Wrong arguments/,
-    '"nodes" with an array containing a non-string or non-number value throws an error.'
-  );
-
-  assert.throws(
-    () => myGraph.addNode(graph.nodes[0]),
-    /The node "n0" already exists./,
-    "Adding an already existing node throws an error."
-  );
-
-  myGraph.addNode({ id: "prototype" }).addNode({ id: "constructor" });
-  assert.ok(
-    myGraph.nodes("prototype") && myGraph.nodes("constructor"),
-    '"constructor" and "prototype" are valid node IDs.'
-  );
-  myGraph.dropNode("prototype").dropNode("constructor");
-
-  // EDGES:
-  // ******
-  assert.deepEqual(
-    graph.edges[0],
-    myGraph.edges(graph.edges[0].id),
-    '"addEdge" works and the edge properties have been preserved.'
-  );
-
-  assert.notStrictEqual(
-    graph.edges[0],
-    myGraph.edges(graph.edges[0].id),
-    'With {clone: true}, "addEdge" creates a new object.'
-  );
-
-  assert.equal(
-    graph.edges[1],
-    myGraph.edges(graph.edges[1].id),
-    'With {clone: false}, "addEdge" keeps the same object.'
-  );
-
-  myGraph.edges(graph.edges[0].id).id = "new_e0";
-  myGraph.edges(graph.edges[0].id).source = "undefined_node";
-  myGraph.edges(graph.edges[0].id).target = "undefined_node";
-  assert.deepEqual(
-    [graph.edges[0].id, graph.edges[0].source, graph.edges[0].target],
-    [
-      myGraph.edges(graph.edges[0].id).id,
-      myGraph.edges(graph.edges[0].id).source,
-      myGraph.edges(graph.edges[0].id).target
-    ],
-    "With {immutable: true}, edge sources, targets and ids in the graph are not writable."
-  );
-
-  const edge = myGraph.edges(graph.edges[1].id);
-  edge.id = "new_e0";
-  edge.source = "undefined_node";
-  edge.target = "undefined_node";
-  assert.deepEqual(
-    ["new_e0", "undefined_node", "undefined_node"],
-    [edge.id, edge.source, edge.target],
-    "With {immutable: false}, edge sources, targets and ids in the graph are writable."
-  );
-  edge.id = "e1";
-  edge.source = "n1";
-  edge.target = "n2";
-
-  myGraph.edges(graph.edges[0].id).myEdgeAttr = 456;
-  assert.strictEqual(
-    456,
-    myGraph.edges(graph.edges[0].id).myEdgeAttr,
-    "Other edge attributes are writable."
-  );
-  myGraph.edges(graph.edges[0].id).myEdgeAttr = 123;
-
-  assert.notStrictEqual(
-    myGraph.edges(),
-    myGraph.edges(),
-    '"edges" without arguments returns a copy of the edge array.'
-  );
-
-  assert.strictEqual(
-    myGraph.edges("unexisting_id"),
-    undefined,
-    '"edges" with an unreferenced id returns undefined and does not throw an error.'
-  );
-
-  assert.deepEqual(
-    myGraph.edges(["e0", "e0"]),
-    [graph.edges[0], graph.edges[0]],
-    '"edges" with a strings array as arguments returns the array of specified edge.'
-  );
-
-  assert.throws(
-    () => myGraph.edges(["e0", {}]),
-
-    /edges: Wrong arguments/,
-    '"edges" with an array containing a non-string or non-number value throws an error.'
-  );
-
-  assert.throws(
-    () => myGraph.addEdge(graph.edges[0]),
-    /The edge "e0" already exists./,
-    "Adding an already existing edge throws an error."
-  );
-
-  // DROPING AND CLEARING:
-  // *********************
-  myGraph.dropNode("n0");
-  assert.deepEqual(
-    myGraph.nodes().map(n => n.id),
-    ["n1", "n2", "n3"],
-    '"dropNode" actually drops the node.'
-  );
-  assert.deepEqual(
-    myGraph.edges().map(e => e.id),
-    ["e1", "e2", "e3", "e4"],
-    '"dropNode" also kills the edges linked to the related nodes..'
-  );
-
-  assert.throws(
-    () => myGraph.dropNode("n0"),
-    /The node "n0" does not exist./,
-    "Droping an unexisting node throws an error."
-  );
-
-  myGraph.dropEdge("e1");
-  assert.deepEqual(
-    myGraph.edges().map(e => e.id),
-    ["e2", "e3", "e4"],
-    '"dropEdge" actually drops the edge.'
-  );
-
-  myGraph.dropEdge("e4");
-  assert.deepEqual(
-    myGraph.edges().map(e => e.id),
-    ["e2", "e3"],
-    '"dropEdge" with a self loops works. (#286)'
-  );
-
-  assert.throws(
-    () => myGraph.dropEdge("e1"),
-    /The edge "e1" does not exist./,
-    "Droping an unexisting edge throws an error."
-  );
-
-  // Reinitialize the graph:
-  myGraph.addNode(graph.nodes[0]);
-  myGraph.addEdge(graph.edges[0]);
-  myGraph.addEdge(graph.edges[1]);
-
-  myGraph.clear();
-  assert.deepEqual(
-    [myGraph.nodes(), myGraph.edges()],
-    [[], []],
-    '"clear" empties the nodes and edges arrays.'
-  );
-
-  myGraph = new sigma.classes.graph();
-  myGraph.read(graph);
-
-  assert.deepEqual(
-    myGraph.nodes(),
-    graph.nodes,
-    '"read" adds properly the nodes.'
-  );
-  assert.deepEqual(
-    myGraph.edges(),
-    graph.edges,
-    '"read" adds properly the edges.'
-  );
-});
-
-QUnit.test("Methods and attached functions", assert => {
-  let counter;
-  const colorPalette = { Person: "#C3CBE1", Place: "#9BDEBD" };
-
-  counter = 0;
-  sigma.classes.graph.attach("addNode", "counterInc", () => counter++);
-
-  sigma.classes.graph.attachBefore("addNode", "applyNodeColorPalette", n => {
-    // eslint-disable-next-line no-param-reassign
-    n.color = colorPalette[n.category];
-  });
-
-  assert.strictEqual(
-    false,
-    sigma.classes.graph.hasMethod("getNodeLabel"),
-    "sigma.classes.hasMethod returns false if the method does not exist."
-  );
-
-  sigma.classes.graph.addMethod("getNodeLabel", function getNodeLabel(nId) {
-    return (this.nodesIndex[nId] || {}).label;
-  });
-
-  assert.strictEqual(
-    true,
-    sigma.classes.graph.hasMethod("getNodeLabel"),
-    "sigma.classes.hasMethod returns true if the method has been added with addMethod."
-  );
-
-  assert.strictEqual(
-    true,
-    sigma.classes.graph.hasMethod("hasMethod"),
-    "sigma.classes.hasMethod returns true if the method is implemented in the core."
-  );
-
-  const myGraph = new sigma.classes.graph();
-  myGraph.addNode({ id: "n0", label: "My node", category: "Person" });
-  assert.strictEqual(
-    1,
-    counter,
-    "Attached functions are effectively executed when the anchor method is called."
-  );
-  assert.strictEqual(
-    myGraph.nodes("n0").color,
-    "#C3CBE1",
-    'Attached "before" functions are effectively executed before when the anchor method is called.'
-  );
-  assert.strictEqual(
-    myGraph.getNodeLabel("n0"),
-    "My node",
-    'Custom methods work, can have arguments, and have access to the data in their scope (through "this").'
-  );
-
-  function noop() {}
-
-  assert.throws(
-    () => sigma.classes.graph.attach("addNode", "counterInc", noop),
-    /A function "counterInc" is already attached to the method "addNode"/,
-    "Attaching a function to a method when there is already a function attached to this method under the same key throws an error."
-  );
-
-  assert.throws(
-    () => sigma.classes.graph.attach("undefinedMethod", "counterInc", noop),
-    /The method "undefinedMethod" does not exist./,
-    "Attaching a function to an unexisting method when throws an error."
-  );
-
-  assert.throws(
-    () =>
-      sigma.classes.graph.attachBefore(
-        "addNode",
-        "applyNodeColorPalette",
-        noop
-      ),
-    /A function "applyNodeColorPalette" is already attached to the method "addNode"/,
-    'Attaching a "before" function to a method when there is already a "before" function attached to this method under the same key throws an error.'
-  );
-
-  assert.throws(
-    () =>
-      sigma.classes.graph.attachBefore(
-        "undefinedMethod",
-        "applyNodeColorPalette",
-        noop
-      ),
-    /The method "undefinedMethod" does not exist./,
-    'Attaching a "before" function to an unexisting method when throws an error.'
-  );
-
-  assert.throws(
-    () => sigma.classes.graph.addMethod("getNodeLabel", noop),
-    /The method "getNodeLabel" already exists./,
-    "Attaching a method whose name is already referenced throws an error."
-  );
-});
-
-QUnit.test("Builtin indexes", assert => {
-  const graph = {
-    nodes: [
-      {
-        id: "n0",
-        label: "Node 0",
-        myNodeAttr: 123
-      },
-      {
-        id: "n1",
-        label: "Node 1"
-      },
-      {
-        id: "n2",
-        label: "Node 2"
-      }
-    ],
-    edges: [
-      {
-        id: "e0",
-        source: "n0",
-        target: "n1",
-        myEdgeAttr: 123
-      },
-      {
-        id: "e1",
-        source: "n1",
-        target: "n2"
-      }
-    ]
-  };
-
-  sigma.classes.graph.addMethod("retrieveIndexes", function getIndexes() {
-    return {
-      inIndex: this.inNeighborsIndex,
-      outIndex: this.outNeighborsIndex,
-      allIndex: this.allNeighborsIndex,
-      inCount: this.inNeighborsCount,
-      outCount: this.outNeighborsCount,
-      allCount: this.allNeighborsCount
-    };
-  });
-
-  const g = new sigma.classes.graph();
-  g.read(graph);
-
-  const index = g.retrieveIndexes();
-
-  assert.deepEqual(
-    index.inIndex,
-    {
-      n0: {},
-      n1: {
-        n0: {
-          e0: {
-            id: "e0",
-            myEdgeAttr: 123,
-            source: "n0",
-            target: "n1"
-          }
-        }
-      },
-      n2: {
-        n1: {
-          e1: {
-            id: "e1",
-            source: "n1",
-            target: "n2"
-          }
-        }
-      }
-    },
-    "Incoming index up to date"
-  );
-
-  assert.deepEqual(
-    index.inCount,
-    {
-      n0: 0,
-      n1: 1,
-      n2: 1
-    },
-    "Incoming count up to date"
-  );
-
-  assert.deepEqual(
-    index.outIndex,
-    {
-      n0: {
-        n1: {
-          e0: {
-            id: "e0",
-            myEdgeAttr: 123,
-            source: "n0",
-            target: "n1"
-          }
-        }
-      },
-      n1: {
-        n2: {
-          e1: {
-            id: "e1",
-            source: "n1",
-            target: "n2"
-          }
-        }
-      },
-      n2: {}
-    },
-    "Outcoming index up to date"
-  );
-
-  assert.deepEqual(
-    index.outCount,
-    {
-      n0: 1,
-      n1: 1,
-      n2: 0
-    },
-    "Outcoming count up to date"
-  );
-
-  assert.deepEqual(
-    index.allIndex,
-    {
-      n0: {
-        n1: {
-          e0: {
-            id: "e0",
-            myEdgeAttr: 123,
-            source: "n0",
-            target: "n1"
-          }
-        }
-      },
-      n1: {
-        n0: {
-          e0: {
-            id: "e0",
-            myEdgeAttr: 123,
-            source: "n0",
-            target: "n1"
-          }
-        },
-        n2: {
-          e1: {
-            id: "e1",
-            source: "n1",
-            target: "n2"
-          }
-        }
-      },
-      n2: {
-        n1: {
-          e1: {
-            id: "e1",
-            source: "n1",
-            target: "n2"
-          }
-        }
-      }
-    },
-    "Full index up to date"
-  );
-
-  assert.deepEqual(
-    index.allCount,
-    {
-      n0: 1,
-      n1: 2,
-      n2: 1
-    },
-    "Full count up to date"
-  );
-
-  g.dropNode("n2");
-
-  assert.deepEqual(
-    index.inIndex,
-    {
-      n0: {},
-      n1: {
-        n0: {
-          e0: {
-            id: "e0",
-            myEdgeAttr: 123,
-            source: "n0",
-            target: "n1"
-          }
-        }
-      }
-    },
-    "Incoming index up to date after having dropped a node"
-  );
-
-  assert.deepEqual(
-    index.inCount,
-    {
-      n0: 0,
-      n1: 1
-    },
-    "Incoming count up to date after having dropped a node"
-  );
-
-  assert.deepEqual(
-    index.outIndex,
-    {
-      n0: {
-        n1: {
-          e0: {
-            id: "e0",
-            myEdgeAttr: 123,
-            source: "n0",
-            target: "n1"
-          }
-        }
-      },
-      n1: {}
-    },
-    "Outcoming index up to date after having dropped a node"
-  );
-
-  assert.deepEqual(
-    index.outCount,
-    {
-      n0: 1,
-      n1: 0
-    },
-    "Outcoming count up to date after having dropped a node"
-  );
-
-  assert.deepEqual(
-    index.allIndex,
-    {
-      n0: {
-        n1: {
-          e0: {
-            id: "e0",
-            myEdgeAttr: 123,
-            source: "n0",
-            target: "n1"
-          }
-        }
-      },
-      n1: {
-        n0: {
-          e0: {
-            id: "e0",
-            myEdgeAttr: 123,
-            source: "n0",
-            target: "n1"
-          }
-        }
-      }
-    },
-    "Full index up to date after having dropped a node"
-  );
-
-  assert.deepEqual(
-    index.allCount,
-    {
-      n0: 1,
-      n1: 1
-    },
-    "Full count up to date after having dropped a node"
-  );
-
-  g.dropEdge("e0");
-
-  assert.deepEqual(
-    index.inIndex,
-    {
-      n0: {},
-      n1: {}
-    },
-    "Incoming index up to date after having dropped an edge"
-  );
-
-  assert.deepEqual(
-    index.inCount,
-    {
-      n0: 0,
-      n1: 0
-    },
-    "Incoming count up to date after having dropped an edge"
-  );
-
-  assert.deepEqual(
-    index.outIndex,
-    {
-      n0: {},
-      n1: {}
-    },
-    "Outcoming index up to date after having dropped an edge"
-  );
-
-  assert.deepEqual(
-    index.outCount,
-    {
-      n0: 0,
-      n1: 0
-    },
-    "Outcoming count up to date after having dropped an edge"
-  );
-
-  assert.deepEqual(
-    index.allIndex,
-    {
-      n0: {},
-      n1: {}
-    },
-    "Full index up to date after having dropped an edge"
-  );
-
-  assert.deepEqual(
-    index.allCount,
-    {
-      n0: 0,
-      n1: 0
-    },
-    "Full count up to date after having dropped an edge"
-  );
-});
-
-QUnit.test("Custom indexes", assert => {
-  sigma.classes.graph.addIndex("nodesCount", {
-    constructor() {
-      this.nodesCount = 0;
-    },
-    addNode() {
-      this.nodesCount++;
-    },
-    dropNode() {
-      this.nodesCount--;
-    },
-    clear() {
-      this.nodesCount = 0;
-    }
-  });
-
-  sigma.classes.graph.addMethod("getNodesCount", function getNodeCount() {
-    return this.nodesCount;
-  });
-
-  const myGraph = new sigma.classes.graph();
-  myGraph
-    .addNode({ id: "n0" })
-    .addNode({ id: "n1" })
-    .dropNode("n0");
-  assert.strictEqual(
-    1,
-    myGraph.getNodesCount(),
-    "Indexes work, and the scope is effectively shared with custom methods."
-  );
-});
